requirements.txt
setuptools-rust
pyyaml
azure-storage-blob